api_gateway/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
PyJWT
